+++ src/gwt/src/org/rstudio/studio/client/application/ui/impl/DesktopApplicationHeader.java
@@ -37,8 +37,7 @@ import org.rstudio.studio.client.application.ui.ApplicationHeader;
 import org.rstudio.studio.client.application.ui.GlobalToolbar;
 import org.rstudio.studio.client.common.GlobalDisplay;
 import org.rstudio.studio.client.common.debugging.ErrorManager;
-import org.rstudio.studio.client.events.BeginPasteEvent;
-import org.rstudio.studio.client.events.EndPasteEvent;
+import org.rstudio.studio.client.events.EditEvent;
 import org.rstudio.studio.client.server.ServerError;
 import org.rstudio.studio.client.server.ServerRequestCallback;
 import org.rstudio.studio.client.workbench.codesearch.CodeSearch;
@@ -59,11 +58,7 @@ import com.google.gwt.core.client.JsArrayString;
 import com.google.gwt.core.client.Scheduler;
 import com.google.gwt.core.client.Scheduler.ScheduledCommand;
 import com.google.gwt.dom.client.Element;
-import com.google.gwt.event.shared.HandlerRegistration;
-import com.google.gwt.user.client.Event;
-import com.google.gwt.user.client.Event.NativePreviewHandler;
 import com.google.gwt.user.client.Timer;
-import com.google.gwt.user.client.Event.NativePreviewEvent;
 import com.google.gwt.user.client.ui.Widget;
 import com.google.inject.Inject;
 import com.google.inject.Provider;
@@ -199,6 +194,11 @@ public class DesktopApplicationHeader implements ApplicationHeader
       toolbar_.focusGoToFunction();
    }
 
+   private void fireEditEvent(final int type)
+   {
+      eventBus_.fireEvent(new EditEvent(true, type));
+   }
+   
    @Handler
    void onUndoDummy()
    {
@@ -214,40 +214,21 @@ public class DesktopApplicationHeader implements ApplicationHeader
    @Handler
    void onCutDummy()
    {
+      fireEditEvent(EditEvent.TYPE_CUT);
       Desktop.getFrame().clipboardCut();
    }
 
    @Handler
    void onCopyDummy()
    {
+      fireEditEvent(EditEvent.TYPE_COPY);
       Desktop.getFrame().clipboardCopy();
    }
 
    @Handler
    void onPasteDummy()
    {
-      eventBus_.fireEvent(new BeginPasteEvent());
-      keyListener_ = Event.addNativePreviewHandler(new NativePreviewHandler()
-      {
-         
-         @Override
-         public void onPreviewNativeEvent(NativePreviewEvent event)
-         {
-            // Defer so that this event can go through and execute unabated
-            Scheduler.get().scheduleDeferred(new ScheduledCommand()
-            {
-               
-               @Override
-               public void execute()
-               {
-                  eventBus_.fireEvent(new EndPasteEvent());
-                  keyListener_.removeHandler();
-                  keyListener_ = null;
-               }
-            });
-         }
-      });
-      
+      fireEditEvent(EditEvent.TYPE_PASTE);
       Desktop.getFrame().clipboardPaste();
    }
 
@@ -406,5 +387,4 @@ public class DesktopApplicationHeader implements ApplicationHeader
    private IgnoredUpdates ignoredUpdates_;
    private boolean ignoredUpdatesDirty_ = false;
    private ApplicationQuit appQuit_; 
-   private HandlerRegistration keyListener_;
 }